+++ Src/SLib/libxml/libxml.h
@@ -2579,7 +2579,7 @@ int xmlNop(void);
 	int    FASTCALL xmlBufGrow(xmlBuf * buf, int len);
 	int    FASTCALL xmlBufInflate(xmlBuf * buf, size_t len);
 	int    FASTCALL xmlBufResize(xmlBuf * buf, size_t len);
-	int    STDCALL  xmlBufAdd(xmlBuf * buf, const xmlChar * str, int len);
+	int    STDCALL  xmlBufAdd(xmlBuf * buf, const xmlChar * str, ssize_t len); // @v12.5.1 (int len)-->(ssize_t len)
 	int    xmlBufAddHead(xmlBuf * buf, const xmlChar * str, int len);
 	int    FASTCALL xmlBufCat(xmlBuf * buf, const xmlChar * str);
 	int    FASTCALL xmlBufCCat(xmlBuf * buf, const char * str);
